requirements.txt
tweepy
streamlit
plotly
pytest
nltk
transformers
torch
textblob
python-dotenv
pandas
wordcloud
vaderSentiment
tweepy>=4.0
python-dotenv